.env issue resolved
